Make scalate url generator support optional, move related tests to scalate
