+++ core/src/test/scala/org/scalatra/UrlGeneratorTestServlet.scala
@@ -1,6 +1,6 @@
 package org.scalatra
 
-class UrlGeneratorTestServlet extends ScalatraServlet with ReverseRoutingSupport
+class UrlGeneratorTestServlet extends ScalatraServlet
 {
   val cat: String = "meea"
